s7-t3 done

+++ src/main/java/s7/Compare.java
@@ -1,12 +1,15 @@
 package s7;
 
+import java.math.BigDecimal;
+
 public class Compare<T extends Number, K extends Number> {
     int compare(T first, K second) {
-        if (first.intValue() > second.intValue()){
+
+        if (first.doubleValue() > second.doubleValue()) {
             return 1;
-        } else if (first.intValue() < second.intValue()) {
+        } else if (first.doubleValue() < second.doubleValue()) {
             return -1;
-        } else{
+        } else {
             return 0;
         }
     }
